requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-mathfilters==1.0.0
et_xmlfile==2.0.0
idna==3.10
openpyxl==3.1.5
pillow==11.2.1
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
reportlab==4.4.2
requests==2.32.3
resend==2.10.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
